refactor(vm): split vm_control_ops.rs into cohesive modules (§7-8)
Split the 3388-line vm_control_ops.rs (over the 500-line limit) into ten
cohesive modules. All but one are under 500 lines; vm_for_loop_body.rs
(686) holds the single exec_for_loop_body function, which cannot be split
at a file boundary without refactoring the function itself.

- vm_control_ops.rs (370)          - loop specs + while loop + scope helpers
- vm_for_loop_dispatch.rs (227)    - exec_for_loop_op
- vm_for_loop_body.rs (686)        - exec_for_loop_body (single mega-fn)
- vm_for_loop_intrange.rs (243)    - exec_for_loop_int_range
- vm_for_loop_lazy.rs (391)        - lazy gather/io-lines for-loop variants
- vm_loop_writeback.rs (225)       - per-element/topic source writeback
- vm_loop_writeback_quant.rs (369) - QuantHash/Hash value writeback + rw param
- vm_loop_cstyle_repeat.rs (210)   - C-style loop + repeat loop
- vm_given_when_ops.rs (302)       - given/when/default
- vm_try_catch_ops.rs (398)        - try/CATCH

All methods are impl Interpreter, so the split is visibility-only. The
three loop-spec structs stay in the main module (pub(super), imported
where used); cross-module private helpers are bumped to pub(super). The
fn multiset is byte-identical. No behavior change.
