+++ src/vm.rs
@@ -144,6 +144,11 @@ mod vm_data_ops;
 mod vm_dispatch_helpers;
 mod vm_env_helpers;
 pub(crate) mod vm_flipflop_ops;
+mod vm_for_loop_body;
+mod vm_for_loop_dispatch;
+mod vm_for_loop_intrange;
+mod vm_for_loop_lazy;
+mod vm_given_when_ops;
 mod vm_helpers;
 mod vm_helpers_junction;
 mod vm_helpers_lazy;
@@ -153,6 +158,9 @@ pub(crate) mod vm_hyper_func;
 mod vm_hyper_method_ops;
 pub(crate) mod vm_hyper_ops;
 mod vm_hyper_race_parallel;
+mod vm_loop_cstyle_repeat;
+mod vm_loop_writeback;
+mod vm_loop_writeback_quant;
 pub(crate) mod vm_meta_ops;
 pub(crate) mod vm_method_dispatch;
 pub(crate) mod vm_misc_assign;
@@ -186,6 +194,7 @@ pub(crate) mod vm_stats;
 pub(crate) mod vm_string_regex_ops;
 pub(crate) mod vm_subst_apply;
 pub(crate) mod vm_subst_exec;
+mod vm_try_catch_ops;
 mod vm_typedecl_ops;
 mod vm_value_helpers;
 mod vm_var_assign_ops;
